Revamp notificationID creation
notificationIDs are now based on a true UNIX time stamp, they are condensed down by removing the first and last two digits of one. This helps the number fit into a java int for notification scheduling and also lets us use the notificationID as a way to see a rough estimate of when a certain reminder was created. Previous users will have to recreate their reminders so that sorting works correctly.